+++ Enums/EnumQuotaType.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace LethalBots.Enums
+{
+    [Serializable]
+    public enum EnumQuotaType
+    {
+        Normal,
+        Fill
+    }
+}